Safe group spawn (#34)
* Adding 'safe' group spawning. Needs massive testing.

* Update garrison max units to fit in the new AI limit. TODO: move to the UI?

* Added some function headers, added HC helper methods to reduce the need for 'piping'

* adding cache funtion headers

* build-a-group comments

* Moved to a single passthrough function

* Refactor garrison, add the rest of the class comments

* Add missing files

* Fix bad inheritance

* Fixing brace match error

* Fix syntax error

* fix some minor errors

* fixing another issue

* more select random syntax updates

* add missing call

* fix missing bracket

* Added new helper function, finished debugging code

* removing debug
