+++ addons/core/ACE_Settings.hpp
@@ -1,0 +1,10 @@
+class ACE_Settings {
+    class GVAR(showNotifcations) {
+        category = QUOTE(PREFIX);
+        displayName = "Potato: Show SystemChat Notifcations";
+        description = "False should disable most Notifcations, can use to make clean recordings.";
+        typeName = "BOOL";
+        value = 1;
+        isClientSettable = 1;
+    };
+};
