+++ addons/reinforceToHC/XEH_PREP.hpp
@@ -2,5 +2,4 @@ TRACE_1("",QUOTE(ADDON));
 
 PREP(reinforcementsCreateUnits);
 PREP(setupReinforcements);
-PREP(setupReinforcementsOnServer);
 PREP(spawnReinforcements);
