+++ addons/reinforceToHC/functions/fnc_setupReinforcements.sqf
@@ -1,3 +1,20 @@
+/*
+ * Author: AACO
+ * Function called by the placed ARES reinforcement module to create the UI
+ * to gather the reinforcment configuration, and send that information to the server
+ *
+ * Arguments:
+ * 0: placed logic object <OBJECT>
+ *
+ * Return Value:
+ * True if the logic is local to execution, false otherwise <BOOL>
+ *
+ * Example:
+ * [gameLogic] call potato_reinforceToHC_fnc_setupReinforcements;
+ *
+ * Public: No
+ */
+
 #include "script_component.hpp"
 
 params["_logic"];
@@ -99,7 +116,7 @@ if (_dialogVehicleClass == UNARMED_HELO_UNIT_POOL_INDEX || _dialogVehicleClass =
 // Choose the LZ based on what the user indicated
 private _lz = switch (_dialogLzAlgorithm) do {
     case 0: { // Random
-        _allLzs call BIS_fnc_selectRandom
+        selectRandom _allLzs
     };
     case 1: { // Nearest
         [_spawnPosition, _allLzs] call Ares_fnc_GetNearest
@@ -108,7 +125,7 @@ private _lz = switch (_dialogLzAlgorithm) do {
         [_spawnPosition, _allLzs] call Ares_fnc_GetFarthest
     };
     case 3: { // Least used
-        private _leastUsed = _allLzs call BIS_fnc_selectRandom; // Choose randomly to start.
+        private _leastUsed = selectRandom _allLzs; // Choose randomly to start.
         {
             if (_x getVariable ["Ares_Lz_Count", 0] < _lz getVariable ["Ares_Lz_Count", 0]) then {
                 _leastUsed = _x;
@@ -143,20 +160,23 @@ if (count (_pool select _vehiclePoolIndex) == 0) exitWith {
     true
 };
 
-//spawn on remote machine
+// send to HC
 [
-    _pool,
-    _vehiclePoolIndex,
-    _spawnPosition,
-    _side,
-    _lz,
-    _lzSize,
-    _dialogVehicleBehaviour,
-    _dialogUnitBehaviour,
-    _allRps,
-    _dialogRpAlgorithm,
-    _rpSize
-] remoteExecCall [QFUNC(setupReinforcementsOnServer), SERVER_CLIENT_ID];
+    [
+        _pool,
+        _vehiclePoolIndex,
+        _spawnPosition,
+        _side,
+        _lz,
+        _lzSize,
+        _dialogVehicleBehaviour,
+        _dialogUnitBehaviour,
+        _allRps,
+        _dialogRpAlgorithm,
+        _rpSize
+    ],
+    QFUNC(spawnReinforcements)
+] call EFUNC(zeusHC,hcPassthrough);
 
 if (count _allRps > 0) then {
     [objNull, "Transport dispatched to LZ. Squad will head to RP."] call bis_fnc_showCuratorFeedbackMessage;
