--- addons/reinforceToHC/functions/fnc_setupReinforcementsOnServer.sqf
@@ -1,6 +1,0 @@
-#include "script_component.hpp"
-
-TRACE_1("params",_this);
-
-//spawn on remote machine
-_this remoteExecCall [QFUNC(spawnReinforcements), [] call EFUNC(zeusHC,getSpawnMachineId)];
