+++ addons/zeusHC/ACE_Settings.hpp
@@ -1,0 +1,10 @@
+class ACE_Settings {
+    class GVAR(maxAICount) {
+        typeName = "NUMBER";
+        value = 90;
+    };
+    class GVAR(maxGroupCountPerSide) {
+        typeName = "NUMBER";
+        value = 140;
+    };
+};
