+++ addons/zeusHC/XEH_PREP.hpp
@@ -4,6 +4,11 @@ PREP(createEntityZeus);
 PREP(createEntityLocal);
 PREP(createEntityDismounts);
 
+PREP(createGroup);
+PREP(getCrewCount);
+PREP(hcPassthrough);
+PREP(canCreateGroup);
+PREP(sendCuratorHint);
 PREP(getSpawnMachineId);
 
 PREP(buildACache);
@@ -11,7 +16,6 @@ PREP(buildAGroupZeus);
 PREP(buildAGroupDialogAdd);
 PREP(buildAGroupSpawnLocal);
 PREP(buildAGroupDialogLoad);
-PREP(buildAGroupSpawnServer);
 PREP(buildAGroupDialogSpawn);
 PREP(buildAGroupDialogRemove);
 PREP(buildAGroupDialogFactionChange);
@@ -20,7 +24,6 @@ PREP(spawnACache);
 PREP(spawnAVicZeus);
 PREP(spawnAVicSpawnLocal);
 PREP(spawnAVicDialogLoad);
-PREP(spawnAVicSpawnServer);
 PREP(spawnAVicDialogSpawn);
 PREP(spawnAVicDialogFactionChange);
 
@@ -32,7 +35,6 @@ PREP(garrisonCache);
 PREP(garrisonLocal);
 PREP(garrisonSpawn);
 PREP(garrisonModule);
-PREP(garrisonServer);
 PREP(garrisonDialogLoad);
 PREP(garrisonDialogSpawn);
 PREP(garrisonIsFacingWall);
