+++ addons/zeusHC/config.cpp
@@ -133,6 +133,7 @@ class CfgPatches {
     };
 };
 
+#include "ACE_Settings.hpp"
 #include "CfgEden.hpp"
 #include "CfgEventHandlers.hpp"
 #include "CfgGarrison.hpp"
