+++ addons/zeusHC/functions/fnc_buildACache.sqf
@@ -1,3 +1,13 @@
+/*
+ * Author: AACO
+ * Function used to create a cache of spawnable units
+ *
+ * Example:
+ * [] call potato_zeusHC_fnc_buildACache;
+ *
+ * Public: No
+ */
+
 #include "script_component.hpp"
 TRACE_1("Params",_this);
