+++ addons/zeusHC/functions/fnc_buildAGroupZeus.sqf
@@ -1,3 +1,16 @@
+/*
+ * Author: AACO
+ * Function called when the build a group module is placed in zeus
+ *
+ * Arguments:
+ * 0: logic object <OBJECT>
+ *
+ * Examples:
+ * [objLogic] call potato_zeusHC_fnc_buildAGroupZeus;
+ *
+ * Public: No
+ */
+
 #include "script_component.hpp"
 TRACE_1("Params",_this);
 
@@ -6,6 +19,5 @@ params ["_logic"];
 if (local _logic) then {
     GVAR(buildSpawnLocation) = getPosATL _logic;
     createDialog QGVAR(build_a_group_dialog);
+    deleteVehicle _logic;
 };
-
-deleteVehicle _logic;
