+++ addons/zeusHC/functions/fnc_garrisonLocal.sqf
@@ -1,29 +1,44 @@
+/*
+ * Author: AACO
+ * Function to spawn a group and have them garrison a building
+ *
+ * Arguments:
+ * 0: Array of unit classnames you want to spawn, should have the same length as the below array <ARRAY>
+ * 1: Array of positions you want to spawn the units at, should have the same length as the above array <ARRAY>
+ * 2: Side of the group you want to spawn <SIDE>
+ *
+ * Example:
+ * [_unitsToAdd, _unitPositions, _side] call potato_zeusHC_fnc_garrisonLocal
+ *
+ * Public: Yes
+ */
+
 #include "script_component.hpp"
 TRACE_1("Params",_this);
 
 _this spawn {
-    params ["_unitsToAdd","_side"];
+    params ["_unitsToAdd","_unitPositions","_side"];
 
     private _building = objNull;
-    private _buildingGroup = createGroup _side;
+    private _buildingGroup = [_side, [0,0,0], _unitsToAdd] call FUNC(createGroup);
 
     if ((isNil "_buildingGroup") || {isNull _buildingGroup}) exitWith {
         ERROR("Group Could not be created");
     };
 
     {
-        private _unit = _buildingGroup createUnit [_x select 0, _x select 1, [], 0, "NONE"];
-        if (isNil "_unit") exitWith {
-            ERROR("Unit Could not be created");
-        };
+        private _unit = _x;
+        private _position = _unitPositions select _forEachIndex;
+        _unit setPos _position;
+
         doStop _unit;
 
         if (isNull _building) then {
-            _building = (nearestObjects [_x select 1, ["house"], 50]) select 0;
+            _building = (nearestObjects [_position, ["house"], 50]) select 0;
         };
 
-        private _inside = [_unit, 0, 0, 25] call FUNC(garrisonIsFacingWall);
-        private _directionToBuilding = [_unit, _building] call BIS_fnc_RelativeDirTo;
+        private _inside = [_unit, 0, 0] call FUNC(garrisonIsFacingWall);
+        private _directionToBuilding = _unit getRelDir _building;
         private _finalDirection = _directionToBuilding - 180;
 
         private _facingWall = if (_inside) then {
@@ -45,7 +60,6 @@ _this spawn {
             };
         };
 
-        _unit doWatch ([_unit, 20, _finalDirection] call BIS_fnc_relPos);
-        nil
-    } count _unitsToAdd;
+        _unit doWatch (_unit getRelPos [20, _finalDirection]);
+    } forEach (units _buildingGroup);
 };
