+++ addons/zeusHC/functions/fnc_garrisonModule.sqf
@@ -1,3 +1,16 @@
+/*
+ * Author: AACO
+ * Function called when the garrison module is placed in zeus or when the garrison module is loaded at init
+ *
+ * Arguments:
+ * 0: logic object <OBJECT>
+ *
+ * Examples:
+ * [objLogic] call potato_zeusHC_fnc_garrisonModule;
+ *
+ * Public: No
+ */
+
 #include "script_component.hpp"
 TRACE_1("Params",_this);
 
@@ -21,6 +34,6 @@ if (local _logic) then {
     } else {
         createDialog QGVAR(garrison_dialog);
     };
-};
 
-deleteVehicle _logic;
+    deleteVehicle _logic;
+};
