+++ addons/zeusHC/functions/fnc_spawnACache.sqf
@@ -1,3 +1,13 @@
+/*
+ * Author: AACO
+ * Function used to create a cache of spawnable vehicles
+ *
+ * Example:
+ * [] call potato_zeusHC_fnc_spawnACache;
+ *
+ * Public: No
+ */
+
 #include "script_component.hpp"
 TRACE_1("Params",_this);
