+++ addons/zeusHC/functions/fnc_spawnAVicDialogFactionChange.sqf
@@ -1,9 +1,25 @@
+/*
+ * Author: AACO
+ * Function used to change the active faction in the dialog
+ * Clear the vehicle list, repopulates the faction vehicle list
+ * Should only be called from UI events
+ *
+ * Arguments:
+ * 0: index of the faction list <NUMBER>
+ *
+ * Examples:
+ * [1] call potato_zeusHC_fnc_spawnAVicDialogFactionChange; //Faction in the first index
+ * [5] call potato_zeusHC_fnc_spawnAVicDialogFactionChange; //Faction in the fifth index
+ *
+ * Public: No
+ */
+
 #include "script_component.hpp"
 TRACE_1("Params",_this);
 
 params [["_index", lbCurSel SPAWN_VIC_FACTIONS_IDC]];
 
-// clear both list boxes
+// clear list box
 lbClear SPAWN_VIC_LIST_IDC;
 
 private _factionIndex = (GVAR(spawnVicCache) select 0) find (lbData [SPAWN_VIC_FACTIONS_IDC, _index]);
