+++ addons/zeusHC/functions/fnc_spawnAVicDialogLoad.sqf
@@ -1,3 +1,11 @@
+/*
+ * Author: AACO
+ * Function used populate the UI, must be spawned
+ * Should only be called from UI events
+ *
+ * Public: No
+ */
+
 #include "script_component.hpp"
 TRACE_1("Params",_this);
