+++ addons/zeusHC/functions/fnc_transferAllToHCZeus.sqf
@@ -1,3 +1,16 @@
+/*
+ * Author: AACO
+ * Function called when the transfer all AI to HC module is placed in zeus
+ *
+ * Arguments:
+ * 0: logic object <OBJECT>
+ *
+ * Examples:
+ * [objLogic] call potato_zeusHC_fnc_transferAllToHCZeus;
+ *
+ * Public: No
+ */
+
 #include "script_component.hpp"
 TRACE_1("Params",_this);
